[Data/AIR] Move `TensorExtension` to `ray.air` for use in other packages (#25517)
Moves Tensor extensions to ray.air to facilitate their use in other Ray libraries (AIR, Serve).
